refactor : enum 패키지 분리 enum class 들을 Constants 패키지로 분리했습니다.
