+++ src/main/java/lotto/Constants/IntConstants.java
@@ -1,11 +1,10 @@
-package lotto;
+package lotto.Constants;
 
 public enum IntConstants {
     MIN_RANGE(1),
     MAX_RANGE(45),
     UNIT_BILL(1000),
     LOTTO_NUMBER_COUNT(6),
-    BONUS_NUMBER_COUNT(1),
     THREE_HIT(3),
     FOUR_HIT(4),
     FIVE_HIT(5),
