+++ src/test/java/lotto/InputViewValidatorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.Constants.ErrorMessage;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
